Schedule - Reorganize data transfer object package
