+++ ScheduleService/src/main/java/rs/raf/gym/controller/TrainingController.java
@@ -22,15 +22,15 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import rs.raf.gym.dto.TrainingCreateDto;
-import rs.raf.gym.dto.TrainingDto;
-import rs.raf.gym.dto.TrainingUpdateDto;
+import rs.raf.gym.dto.training.TrainingCreateDto;
+import rs.raf.gym.dto.training.TrainingDto;
+import rs.raf.gym.dto.training.TrainingUpdateDto;
 import rs.raf.gym.model.TrainingType;
 import rs.raf.gym.service.ITrainingService;
 
@@ -56,18 +56,18 @@ public class TrainingController {
         return new ResponseEntity<>(trainingService.findAll(pageable), HttpStatus.OK);
     }
 
-    @GetMapping("/type={type}")
-    public ResponseEntity<List<TrainingDto>> getAllTrainingsWithType(@PathVariable("type") @Valid TrainingType trainingType) {
+    @GetMapping(params = "type")
+    public ResponseEntity<List<TrainingDto>> getAllTrainingsWithType(@RequestParam("type") @Valid TrainingType trainingType) {
         return new ResponseEntity<>(trainingService.findByType(trainingType), HttpStatus.OK);
     }
 
-    @GetMapping(value = "/type={type}", params = { "page", "size" })
-    public ResponseEntity<Page<TrainingDto>> getAllTrainingsWithType(@PathVariable("type") @Valid TrainingType trainingType, Pageable pageable) {
+    @GetMapping(params = { "page", "size", "type" })
+    public ResponseEntity<Page<TrainingDto>> getAllTrainingsWithType(@RequestParam("type") @Valid TrainingType trainingType, Pageable pageable) {
         return new ResponseEntity<>(trainingService.findByType(trainingType, pageable), HttpStatus.OK);
     }
 
-    @GetMapping("/name={name}")
-    public ResponseEntity<TrainingDto> getTrainingWithName(@PathVariable("name") String name) {
+    @GetMapping(params = "name")
+    public ResponseEntity<TrainingDto> getTrainingWithName(@RequestParam("name") String name) {
         return new ResponseEntity<>(trainingService.findByName(name), HttpStatus.OK);
     }
