+++ ScheduleService/src/main/java/rs/raf/gym/dto/gym/GymUpdateDto.java
@@ -14,15 +14,29 @@
  * limitations under the License.
  */
 
-package rs.raf.gym.dto;
+package rs.raf.gym.dto.gym;
 
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.PositiveOrZero;
 import lombok.Getter;
 import lombok.Setter;
 
 @Setter
 @Getter
-public class TrainingTypeDto {
+public class GymUpdateDto {
 
+    @NotNull
+    private String oldName;
+
+    @NotNull
     private String name;
+
+    private String description;
+
+    private Long managerId;
+
+    @NotNull
+    @PositiveOrZero
+    private Integer trainers;
 
 }
