+++ ScheduleService/src/main/java/rs/raf/gym/service/implementation/GymService.java
@@ -19,9 +19,9 @@ package rs.raf.gym.service.implementation;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import rs.raf.gym.dto.GymCreateDto;
-import rs.raf.gym.dto.GymDto;
-import rs.raf.gym.dto.GymUpdateDto;
+import rs.raf.gym.dto.gym.GymCreateDto;
+import rs.raf.gym.dto.gym.GymDto;
+import rs.raf.gym.dto.gym.GymUpdateDto;
 import rs.raf.gym.mapper.GymMapper;
 import rs.raf.gym.model.Gym;
 import rs.raf.gym.repository.IGymRepository;
