+++ ScheduleService/src/main/java/rs/raf/gym/service/implementation/TrainingService.java
@@ -19,9 +19,9 @@ package rs.raf.gym.service.implementation;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import rs.raf.gym.dto.TrainingCreateDto;
-import rs.raf.gym.dto.TrainingDto;
-import rs.raf.gym.dto.TrainingUpdateDto;
+import rs.raf.gym.dto.training.TrainingCreateDto;
+import rs.raf.gym.dto.training.TrainingDto;
+import rs.raf.gym.dto.training.TrainingUpdateDto;
 import rs.raf.gym.mapper.TrainingMapper;
 import rs.raf.gym.model.Training;
 import rs.raf.gym.model.TrainingType;
