+++ ScheduleService/src/main/java/rs/raf/gym/service/implementation/TrainingTypeService.java
@@ -19,9 +19,9 @@ package rs.raf.gym.service.implementation;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import rs.raf.gym.dto.TrainingTypeCreateDto;
-import rs.raf.gym.dto.TrainingTypeDto;
-import rs.raf.gym.dto.TrainingTypeUpdateDto;
+import rs.raf.gym.dto.training_type.TrainingTypeCreateDto;
+import rs.raf.gym.dto.training_type.TrainingTypeDto;
+import rs.raf.gym.dto.training_type.TrainingTypeUpdateDto;
 import rs.raf.gym.mapper.TrainingTypeMapper;
 import rs.raf.gym.model.TrainingType;
 import rs.raf.gym.repository.ITrainingTypeRepository;
